fix: broken toml snippets
