+++ pages/blog/[year]/[month]/[day]/[slug].tsx
@@ -3,7 +3,6 @@ import { serialize } from 'next-mdx-remote/serialize'
 import { NextSeo } from 'next-seo'
 import Image from 'next/image'
 import { useRouter } from 'next/router'
-import { useMemo } from 'react'
 import { generateReadingTime } from 'lib/helpers'
 import { getAllPostSlugs, getPostData, getSortedPosts, Post } from 'lib/blog/posts'
 import { MDXRemote, MDXRemoteProps } from 'next-mdx-remote'
@@ -20,8 +19,8 @@ import Link from 'next/link'
 import clsx from 'clsx'
 import { BlogHeader, BlogPrevNext, BlogSidebar, CallToAction } from 'components/sections'
 import { LinkedInLogo, Logo, TwitterLogo } from 'components/svgs'
-import { CopyButton } from 'components/elements'
 import { TwitterTweetEmbed } from 'react-twitter-embed'
+import { Pre } from 'components/blog/Pre'
 
 export async function getStaticPaths() {
 	const paths = getAllPostSlugs()
@@ -104,76 +103,6 @@ export async function getStaticProps({
 	}
 }
 
-const Pre = ({ children, ...props }: any) => {
-	let line = 1
-
-	const code = useMemo(() => {
-		return [children.props.children]
-			.flat()
-			.flatMap((child) => {
-				if (typeof child !== 'string') {
-					return child.props.children
-				} else {
-					return child
-				}
-			})
-			.join('')
-	}, [children])
-
-	return (
-		<div className={clsx('relative')}>
-			<CopyButton code={code} className='absolute right-2 top-2 inline-flex items-center' />
-			<pre
-				{...props}
-				className={clsx(
-					'!border !border-black/10 !bg-white !pr-16 !text-sm text-[#525151] dark:!border-white/10 dark:!bg-black dark:text-[#7A7A7A] [&>*]:!bg-white dark:[&>*]:!bg-black',
-					props.className ?? 'language-'
-				)}
-			>
-				{{
-					...children,
-					props: {
-						...children.props,
-						className: children.props.className ?? 'language-',
-						children: [
-							// <span
-							// 	className='mr-4 inline-block w-4 select-none text-right italic text-[rgb(92,99,112)] last:hidden'
-							// 	key={line}
-							// >
-							// 	{line}
-							// </span>,
-							...[children.props.children].flat().flatMap((child) => {
-								if (typeof child === 'string') {
-									const [head, ...tail] = child.split('\n')
-									return [
-										head,
-										...tail.flatMap((child) => {
-											line++
-
-											return [
-												'\n',
-												// <span
-												// 	key={head}
-												// 	className='mr-4 inline-block w-4 select-none text-right italic text-[rgb(92,99,112)] last:hidden'
-												// >
-												// 	{line}
-												// </span>,
-												child,
-											]
-										}),
-									]
-								} else {
-									return child
-								}
-							}),
-						],
-					},
-				}}
-			</pre>
-		</div>
-	)
-}
-
 const mdxComponents: MDXRemoteProps['components'] = {
 	a(props) {
 		if (props?.href?.match(/^https?:\/\//)) {
@@ -187,17 +116,17 @@ const mdxComponents: MDXRemoteProps['components'] = {
 
 		return <Link {...(props as any)} className='my-0 no-underline'></Link>
 	},
-	pre: (props: any) => {
+	pre: (props) => {
 		return <Pre {...props} />
 	},
-	Tweet: (props: any) => {
+	Tweet: (props) => {
 		return (
 			<div className='flex items-center justify-center [&>*]:w-full [&>*]:max-w-sm lg:[&>*]:max-w-xl'>
 				<TwitterTweetEmbed {...props} options={{ width: '100%' }} />
 			</div>
 		)
 	},
-	TLDR: (props: any) => {
+	TLDR: (props) => {
 		return (
 			<div className='mb-24 text-left text-xl'>
 				<span className='font-bold text-black dark:text-[#C2C2C2]'>TLDR;</span>
@@ -205,7 +134,7 @@ const mdxComponents: MDXRemoteProps['components'] = {
 			</div>
 		)
 	},
-	CaptionedImage: (props: any) => {
+	CaptionedImage: (props) => {
 		return (
 			<div className='relative grid w-full grid-cols-1 justify-items-center'>
 				<Image
